+++ addon/components/bs-button.js
@@ -2,6 +2,8 @@ import Ember from 'ember';
 import TypeClass from 'ember-bootstrap/mixins/type-class';
 import SizeClass from 'ember-bootstrap/mixins/size-class';
 import I18nSupport from 'ember-bootstrap/mixins/i18n-support';
+import ComponentChild from 'ember-bootstrap/mixins/component-child';
+
 
 /**
  Implements a HTML button element, with support for all [Bootstrap button CSS styles](http://getbootstrap.com/css/#buttons)
@@ -82,7 +84,7 @@ import I18nSupport from 'ember-bootstrap/mixins/i18n-support';
  @uses Mixins.SizeClass
  @uses Mixins.I18nSupport
 */
-export default Ember.Component.extend(TypeClass, SizeClass, I18nSupport, {
+export default Ember.Component.extend(ComponentChild, TypeClass, SizeClass, I18nSupport, {
     tagName: 'button',
     classNames: ['btn'],
     classNameBindings: ['active', 'block:btn-block'],
@@ -95,7 +97,7 @@ export default Ember.Component.extend(TypeClass, SizeClass, I18nSupport, {
      */
     classTypePrefix: 'btn',
 
-    attributeBindings: ['id', 'disabled', 'buttonType:type'],
+    attributeBindings: ['disabled', 'buttonType:type'],
 
     /**
      * Default label of the button. Not need if used as a block component
@@ -235,9 +237,12 @@ export default Ember.Component.extend(TypeClass, SizeClass, I18nSupport, {
     },
 
     resetObserver: Ember.observer('reset', function(){
-        if(this.get('reset')){
-            this.resetState();
-        }
+      if(this.get('reset')){
+        Ember.run.scheduleOnce('actions', this, function() {
+          this.set('textState', 'default');
+        });
+
+      }
     }),
 
     text: Ember.computed('textState', 'defaultText', 'pendingText', 'resolvedText', 'rejectedText', function() {
@@ -277,7 +282,11 @@ export default Ember.Component.extend(TypeClass, SizeClass, I18nSupport, {
             }
         };
         this.sendAction('action', this.get('value'), evt, callback);
-    }
+    },
 
+  init: function() {
+    this._super();
+    this.get('reset');
+  }
 
 });